more work

+++ Slicer/SlicerLayer.cs
@@ -54,20 +54,20 @@ namespace MatterHackers.MatterSlice
 		public List<SlicePerimeterSegment> SegmentList = new List<SlicePerimeterSegment>();
 
 		private Polygons openPolygonList = new Polygons();
-		private int z;
+		private long z;
 
-		public MeshProcessingLayer(int z)
+		public MeshProcessingLayer(long z)
 		{
 			this.z = z;
 		}
 
-		public MeshProcessingLayer(int z, string segmentListData)
+		public MeshProcessingLayer(long z, string segmentListData)
 		{
 			this.z = z;
 			SegmentList = CreateSegmentListFromString(segmentListData);
 		}
 
-		public int Z { get { return z; } }
+		public long Z { get { return z; } }
 
 		public static List<SlicePerimeterSegment> CreateSegmentListFromString(string segmentListData)
 		{
